PLANNER-580: Make org.optaplanner.core.api.score.Score class hierarchy JavaScript-compile friendly
