+++ optaplanner-benchmark/src/main/java/org/optaplanner/benchmark/config/blueprint/SolverBenchmarkBluePrintType.java
@@ -22,6 +22,7 @@ import java.util.List;
 import org.optaplanner.benchmark.config.SolverBenchmarkConfig;
 import org.optaplanner.core.config.constructionheuristic.ConstructionHeuristicPhaseConfig;
 import org.optaplanner.core.config.constructionheuristic.ConstructionHeuristicType;
+import org.optaplanner.core.config.constructionheuristic.ConstructionHeuristicTypeHelper;
 import org.optaplanner.core.config.localsearch.LocalSearchPhaseConfig;
 import org.optaplanner.core.config.localsearch.LocalSearchType;
 import org.optaplanner.core.config.phase.PhaseConfig;
@@ -61,7 +62,7 @@ public enum SolverBenchmarkBluePrintType {
     }
 
     private List<SolverBenchmarkConfig> buildEveryConstructionHeuristicType() {
-        ConstructionHeuristicType[] chTypes = ConstructionHeuristicType.getBluePrintTypes();
+        ConstructionHeuristicType[] chTypes = ConstructionHeuristicTypeHelper.getBluePrintTypes();
         List<SolverBenchmarkConfig> solverBenchmarkConfigList = new ArrayList<>(chTypes.length);
         for (ConstructionHeuristicType chType : chTypes) {
             solverBenchmarkConfigList.add(buildSolverBenchmarkConfig(chType, null));
@@ -79,7 +80,7 @@ public enum SolverBenchmarkBluePrintType {
     }
 
     private List<SolverBenchmarkConfig> buildEveryConstructionHeuristicTypeWithEveryLocalSearchType() {
-        ConstructionHeuristicType[] chTypes = ConstructionHeuristicType.getBluePrintTypes();
+        ConstructionHeuristicType[] chTypes = ConstructionHeuristicTypeHelper.getBluePrintTypes();
         LocalSearchType[] lsTypes = LocalSearchType.getBluePrintTypes();
         List<SolverBenchmarkConfig> solverBenchmarkConfigList = new ArrayList<>(
                 chTypes.length * lsTypes.length);
