+++ optaplanner-core/src/main/java/org/optaplanner/core/config/constructionheuristic/ConstructionHeuristicType.java
@@ -16,13 +16,6 @@
 
 package org.optaplanner.core.config.constructionheuristic;
 
-import org.optaplanner.core.config.constructionheuristic.placer.EntityPlacerConfig;
-import org.optaplanner.core.config.constructionheuristic.placer.PooledEntityPlacerConfig;
-import org.optaplanner.core.config.constructionheuristic.placer.QueuedEntityPlacerConfig;
-import org.optaplanner.core.config.constructionheuristic.placer.QueuedValuePlacerConfig;
-import org.optaplanner.core.config.heuristic.selector.entity.EntitySorterManner;
-import org.optaplanner.core.config.heuristic.selector.value.ValueSorterManner;
-
 public enum ConstructionHeuristicType {
     /**
      * A specific form of {@link #ALLOCATE_ENTITY_FROM_QUEUE}.
@@ -70,81 +63,4 @@ public enum ConstructionHeuristicType {
      * Repeat until all entities are assigned.
      */
     ALLOCATE_FROM_POOL;
-
-    public EntitySorterManner getDefaultEntitySorterManner() {
-        switch (this) {
-            case FIRST_FIT:
-            case WEAKEST_FIT:
-            case STRONGEST_FIT:
-                return EntitySorterManner.NONE;
-            case FIRST_FIT_DECREASING:
-            case WEAKEST_FIT_DECREASING:
-            case STRONGEST_FIT_DECREASING:
-                return EntitySorterManner.DECREASING_DIFFICULTY;
-            case ALLOCATE_ENTITY_FROM_QUEUE:
-            case ALLOCATE_TO_VALUE_FROM_QUEUE:
-            case CHEAPEST_INSERTION:
-            case ALLOCATE_FROM_POOL:
-                return EntitySorterManner.DECREASING_DIFFICULTY_IF_AVAILABLE;
-            default:
-                throw new IllegalStateException("The constructionHeuristicType (" + this + ") is not implemented.");
-        }
-    }
-
-    public ValueSorterManner getDefaultValueSorterManner() {
-        switch (this) {
-            case FIRST_FIT:
-            case FIRST_FIT_DECREASING:
-                return ValueSorterManner.NONE;
-            case WEAKEST_FIT:
-            case WEAKEST_FIT_DECREASING:
-                return ValueSorterManner.INCREASING_STRENGTH;
-            case STRONGEST_FIT:
-            case STRONGEST_FIT_DECREASING:
-                return ValueSorterManner.DECREASING_STRENGTH;
-            case ALLOCATE_ENTITY_FROM_QUEUE:
-            case ALLOCATE_TO_VALUE_FROM_QUEUE:
-            case CHEAPEST_INSERTION:
-            case ALLOCATE_FROM_POOL:
-                return ValueSorterManner.INCREASING_STRENGTH_IF_AVAILABLE;
-            default:
-                throw new IllegalStateException("The constructionHeuristicType (" + this + ") is not implemented.");
-        }
-    }
-
-    public EntityPlacerConfig newEntityPlacerConfig() {
-        switch (this) {
-            case FIRST_FIT:
-            case FIRST_FIT_DECREASING:
-            case WEAKEST_FIT:
-            case WEAKEST_FIT_DECREASING:
-            case STRONGEST_FIT:
-            case STRONGEST_FIT_DECREASING:
-            case ALLOCATE_ENTITY_FROM_QUEUE:
-                return new QueuedEntityPlacerConfig();
-            case ALLOCATE_TO_VALUE_FROM_QUEUE:
-                return new QueuedValuePlacerConfig();
-            case CHEAPEST_INSERTION:
-            case ALLOCATE_FROM_POOL:
-                return new PooledEntityPlacerConfig();
-            default:
-                throw new IllegalStateException("The constructionHeuristicType (" + this + ") is not implemented.");
-        }
-    }
-
-    /**
-     * @return {@link #values()} without duplicates (abstract types that end up behaving as one of the other types).
-     */
-    public static ConstructionHeuristicType[] getBluePrintTypes() {
-        return new ConstructionHeuristicType[] {
-            FIRST_FIT,
-            FIRST_FIT_DECREASING,
-            WEAKEST_FIT,
-            WEAKEST_FIT_DECREASING,
-            STRONGEST_FIT,
-            STRONGEST_FIT_DECREASING,
-            CHEAPEST_INSERTION
-        };
-    }
-
 }
