+++ optaplanner-core/src/main/java/org/optaplanner/core/config/heuristic/selector/value/ValueSorterManner.java
@@ -17,8 +17,6 @@
 package org.optaplanner.core.config.heuristic.selector.value;
 
 import org.optaplanner.core.api.domain.variable.PlanningVariable;
-import org.optaplanner.core.impl.domain.variable.descriptor.GenuineVariableDescriptor;
-import org.optaplanner.core.impl.heuristic.selector.common.decorator.SelectionSorter;
 
 /**
  * The manner of sorting a values for a {@link PlanningVariable}.
@@ -29,49 +27,4 @@ public enum ValueSorterManner {
     INCREASING_STRENGTH_IF_AVAILABLE,
     DECREASING_STRENGTH,
     DECREASING_STRENGTH_IF_AVAILABLE;
-
-    public boolean hasSorter(GenuineVariableDescriptor variableDescriptor) {
-        switch (this) {
-            case NONE:
-                return false;
-            case INCREASING_STRENGTH:
-            case DECREASING_STRENGTH:
-                return true;
-            case INCREASING_STRENGTH_IF_AVAILABLE:
-                return variableDescriptor.getIncreasingStrengthSorter() != null;
-            case DECREASING_STRENGTH_IF_AVAILABLE:
-                return variableDescriptor.getDecreasingStrengthSorter() != null;
-            default:
-                throw new IllegalStateException("The sorterManner ("
-                        + this + ") is not implemented.");
-        }
-    }
-
-    public SelectionSorter determineSorter(GenuineVariableDescriptor variableDescriptor) {
-        SelectionSorter sorter;
-        switch (this) {
-            case NONE:
-                throw new IllegalStateException("Impossible state: hasSorter() should have returned null.");
-            case INCREASING_STRENGTH:
-            case INCREASING_STRENGTH_IF_AVAILABLE:
-                sorter = variableDescriptor.getIncreasingStrengthSorter();
-                break;
-            case DECREASING_STRENGTH:
-            case DECREASING_STRENGTH_IF_AVAILABLE:
-                sorter = variableDescriptor.getDecreasingStrengthSorter();
-                break;
-            default:
-                throw new IllegalStateException("The sorterManner ("
-                        + this + ") is not implemented.");
-        }
-        if (sorter == null) {
-            throw new IllegalArgumentException("The sorterManner (" + this
-                    + ") on entity class (" + variableDescriptor.getEntityDescriptor().getEntityClass()
-                    + ")'s variable (" + variableDescriptor.getVariableName()
-                    + ") fails because that variable getter's " + PlanningVariable.class.getSimpleName()
-                    + " annotation does not declare any strength comparison.");
-        }
-        return sorter;
-    }
-
 }
